ref(setup): collapse into one smart view
