+++ spot-client/src/spot-tv/ui/components/setup/index.js
@@ -1,3 +1,4 @@
+export { default as CalendarAuth } from './calendar-auth';
 export { default as Profile } from './profile';
 export { SelectMedia } from './select-media';
-export { default as Setup } from './setup';
+export { default as SelectRoom } from './select-room';
